Small Update
- Tried a couple of different things to get the movement to be
water-like, including using a physics volume, but it doesn't seem to let
me control the jumps when the character is "swimming".  So I have just
some tweaks to the character movement component that make it feel a
little bit like it's in water. Also couldn't get it to be able to have
multiple jumps.

Google has been no help to me...